Formatted code as per drupal coding standards
