+++ jqplot.libraries.yml
@@ -1,11 +1,3 @@
-jqplot.example:
-  version: VERSION
-  js:
-    js/jqplot_example.js: {}
-  dependencies:
-    - core/jquery
-    - core/drupal
-
 jqplot:
   title: jQPlot charting library
   version: 1.0.8
@@ -16,10 +8,8 @@ jqplot:
     gpl-compatible: true
   js:
     /libraries/jquery.jqplot/jquery.jqplot.min.js: { weight: 0, group: JS_LIBRARY}
-    /libraries/jquery.jqplot/plugins/jqplot.pointLabels.min.js: { weight: 0, group: JS_LIBRARY}
   css:
     theme:
-      /libraries/jquery.jqplot/jquery.jqplot.css: {}
       /libraries/jquery.jqplot/jquery.jqplot.min.css: {}
 
 jqplot.json2:
@@ -48,6 +38,7 @@ jqplot.barRenderer:
     /libraries/jquery.jqplot/plugins/jqplot.barRenderer.min.js: { weight: 0, group: JS_LIBRARY}
   dependencies:
     - core/jquery
+    - jqplot/jqplot
 
 jqplot.BezierCurveRenderer:
   title: jQPlot charting library
