--- jqplot_example/jqplot_example.module
@@ -1,5 +1,0 @@
-<?php
-
-/**
- * @file
- */
