+++ jqplot_example/jqplot_example.routing.yml
@@ -2,7 +2,7 @@ jqplot.example:
   path: 'jqplot_example'
   defaults:
     _title: 'jqPlot Charts Library Example'
-    _controller: '\Drupal\jqplot_example\Controller\jqplotExampleController::info'
+    _controller: '\Drupal\jqplot_example\Controller\JqplotExampleController::info'
   requirements:
     _access: 'TRUE'
 
@@ -10,7 +10,7 @@ jqplot.lineCharts:
   path: 'jqplot_example/line_charts'
   defaults:
     _title: 'jqPlot Charts Library Example for line Charts'
-    _controller: '\Drupal\jqplot_example\Controller\jqplotExampleController::lineCharts'
+    _controller: '\Drupal\jqplot_example\Controller\JqplotExampleController::lineCharts'
   requirements:
     _access: 'TRUE'
 
@@ -18,7 +18,7 @@ jqplot.BarCharts:
   path: 'jqplot_example/bar_charts'
   defaults:
     _title: 'jqPlot Charts Library Example for Bar Charts'
-    _controller: '\Drupal\jqplot_example\Controller\jqplotExampleController::BarCharts'
+    _controller: '\Drupal\jqplot_example\Controller\JqplotExampleController::barCharts'
   requirements:
     _access: 'TRUE'
 
@@ -26,7 +26,7 @@ jqplot.PieCharts:
   path: 'jqplot_example/pie_charts'
   defaults:
     _title: 'jqPlot Charts Library Example for Pie Charts'
-    _controller: '\Drupal\jqplot_example\Controller\jqplotExampleController::PieCharts'
+    _controller: '\Drupal\jqplot_example\Controller\JqplotExampleController::pieCharts'
   requirements:
     _access: 'TRUE'
